requirements.txt
google-api-python-client==2.149.0
google-auth==2.35.0
openai>=1.50.0
jinja2>=3.1.4
typer[all]>=0.12.5
rich>=13.8.1
python-dotenv>=1.0.1
pymupdf>=1.24.9
 pypdf==6.2.0
pdfplumber>=0.11.8

